debug and rules organization.
