+++ init.c
@@ -16,7 +16,7 @@ void initRandomNoise(Grid *g){
 }
 
 /* sets cells to 1 with probability "threshold", to 0 otherwise */
-void initRandomBool(Grid *g, float threshold){
+int initRandomBool(Grid *g, float threshold){
   int i, j;
   Cell tempCell;
 
@@ -28,10 +28,12 @@ void initRandomBool(Grid *g, float threshold){
       }
     }
     commitGridUpdate(g);
+		return 1;
   }
 
   else {
-    printf("Hai scelto un valore di probabilità non valido per l'inizializzazione. Scegli un decimale tra 0 e 1");
+    printf("Hai scelto un valore di probabilità non valido per l'inizializzazione. Scegli un decimale tra 0 e 1\n");
+		return 0;
   }
 
 }
@@ -49,4 +51,24 @@ void initBlank(Grid *g, float state){
   }
 
   commitGridUpdate(g);
+}
+
+void initInverseSquare(Grid *g) {
+	int i, j;
+	Cell tempCell;
+	float centerX = (float) (g->width-1) / 2;
+	float centerY = (float) (g->height-1) / 2;
+	float temp;
+
+	for ( i = 0; i < g->width; i++) {
+		for (j = 0; j < g->height; j++) {
+			temp = (centerX-i)*(centerX-i) + (centerY-j)*(centerY-j);
+			/* hack to make it work nicely on large grids: */
+			temp /= centerX*centerX+centerY*centerY;
+			tempCell.data = 1/(temp+1);
+			setCell(g, i, j, tempCell);
+		}
+	}
+
+	commitGridUpdate(g);
 }
