+++ rules.h
@@ -16,16 +16,10 @@ void applyRuleNormalize(Grid *g, double tgtMin, double tgtMax);
 
 void applyRuleSetMass(Grid *g, double tgtMass);
 
-void initRuleConway(Grid *tgt);
-
 void slideshowRuleConvolve(Grid *g, const Grid *op, const char *filename);
 
 void slideshowRuleConway(Grid *g, const char *filename);
 
 void printGrid(const Grid *g);
-
-void initRandomGrid(Grid *g);
-
-void initInverseSquare(Grid *g);
 
 #endif /*RULES_H_INCLUDED*/
